+++ packages/hooks/index.ts
@@ -1,1 +1,4 @@
 export { default as useInView } from "./use-in-view";
+export { default as useInfiniteScroll } from "./use-infinite-scroll";
+export { default as useArchiveInfiniteScroll } from "./use-archive-infinite-scroll";
+export { default as usePostTypeInfiniteScroll } from "./use-post-type-infinite-scroll";