+++ packages/core/CHANGELOG.md
@@ -117,6 +117,13 @@
   - @frontity/connect@1.1.2
   - @frontity/file-settings@1.1.6
 
+## 1.7.1-infinite-scroll-beta.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @frontity/file-settings@1.1.5-infinite-scroll-beta.0
+
 ## 1.7.0
 
 ### Minor Changes