+++ packages/mars-theme/CHANGELOG.md
@@ -39,6 +39,15 @@
   - frontity@1.9.0
   - @frontity/html2react@1.3.4
 
+## 1.4.3-infinite-scroll-beta.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - frontity@1.8.1-infinite-scroll-beta.0
+  - @frontity/components@1.3.3-infinite-scroll-beta.0
+  - @frontity/html2react@1.3.4-infinite-scroll-beta.0
+
 ## 1.4.2
 
 ### Patch Changes